requirements.txt
pandas
bs4
sqlalchemy
psycopg2-binary
curl_cffi
brotli
requests
cloudscraper
dbt-postgres
dbt-duckdb
duckdb